+++ chain/contracts/OptimizedFruitStandV2.sol
@@ -445,7 +445,7 @@ contract OptimizedFruitStandV2 {
 
     function payout(address user, UserStake memory stake)
         internal
-        returns (uint8 errCode)
+        returns (uint256 errCode)
     {
         uint256 blockDelta = block.number - stake.startBlock;
         if (blockDelta > 300) {
