+++ chain/contracts/OptimizedFruitStandV3.sol
@@ -55,7 +55,7 @@ contract OptimizedFruitStandV3 {
 
     function payout(address user, UserStake memory stake)
         internal
-        returns (uint8 errCode)
+        returns (uint256 errCode)
     {
         uint256 blockDelta = block.number - stake.startBlock;
         if (blockDelta > 300) {
